requirements.txt
Flask==2.3.3
google-generativeai  # For GPT model access